feat: add endpoint to REST API to return data about the number of potential donors for each recipient blood type. Also, some unnecessary comments was removed.

+++ backend/src/main/java/com/ronilsonalves/bancodesangue/data/dto/IdadeMediaPorTipoSanguineoDto.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @AllArgsConstructor
 public class IdadeMediaPorTipoSanguineoDto {
     private String tipoSanguineo;
-    private double idadeMedia;
+    private int idadeMedia;
 }
